\name{rRACES-package}
\alias{rRACES}
\alias{rRACES-package}
\docType{package}
\title{
    rRACES: An R wrapper for RACES (RACES is an Advanced Cancer Evolution Simulator)
}
\description{
    This package is an R wrapper for RACES (https://github.com/albertocasagrande/RACES)
    and it exposes RACES classes to R.  
}
\author{
Giulio Caravagna
Alberto Casagrande

\strong{Maintainer:} Giulio Caravagna <gcaravagna@units.it>
}
\keyword{ package }
\concept{cancer evolution}
\concept{read simulation}
\concept{phylogenetic tree}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
